requirements.txt
pandas>=1.5.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
census>=0.8.19
python-dotenv>=1.0.0